fix(daemon): narrow perf server test API
